pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["tyropro <tyroprocodes@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = {extras = ["voice"], version = "^2.3.2"}
pyyaml = "^6.0.1"
requests = "^2.31.0"
yt-dlp = "^2024.3.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
